requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.1
python-dotenv==1.1.0
Werkzeug==3.1.3
sqlalchemy==2.0.40
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-jose==3.3.0
bcrypt==4.1.2 